pyproject.toml
[tool.poetry]
name = "sales sage"
version = "0.1.0"
description = ""
authors = ["M Shahzaib <shahzaibchenna@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.6"
pandas = "^2.2.2"
python-decouple = "^3.8"
langchain-community = "^0.2.0"
langchain-openai = "^0.1.7"
langchain = "^0.2.0"
openai = "^1.30.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
